use a HashMap for the outer cdx map
we only care about the timestamps being ordered, it does not matter for
urls, and we're way past the point of enough items for HashMap to be
faster than a BTreeMap. memory usage seems about the same too.

hopefully this will make search results more interesting, instead of
grouping all the similar urls together