redirect unknown urls based on referer
i was browsing the wayback machine and noticed that pktz.fr's xml
documents, which expect an xslt stylesheet at
/_stylesheet/atom_stylesheet.xsl, rendered properly despite WBM not
rewriting that url. turns out they do some fancy redirecting based on
the referer (sic) header! lets do that too

this lets relative-to-the-root urls mostly work for formats we cannot
rewrite the urls in